+++ Bamboo/Resources/Network/Constants.swift
@@ -6,7 +6,7 @@
 //
 
 // MARK: - API Url들
-public let baseAPI: String = "http://13.209.33.18:8080"
+public let baseAPI: String = "https://server.v1.daesoup.com"
 public let userAPI: String = "\(baseAPI)/user"
 public let postAPI: String = "\(baseAPI)/post"
 public let commentAPI: String = "\(baseAPI)/comment"
