+++ Bamboo/Sources/View/Search/SearchView.swift
