+++ Bamboo/Sources/View/Splash/SplashView.swift
